# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World UZ",
        "description": "Documentation for World Uz Company",
        "contact": {
            "email": "rustamovramziddin7@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in user ",
                "description": "Login by email, password !",
                "operationId": "3baa3b145384161275c3581002c9225f",
                "requestBody": {
                    "description": "Pass Balance history  credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "string",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "text",
                                        "format": "string",
                                        "example": "1234567890AA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register ",
                "description": "Register user !",
                "operationId": "9f2c210e9e11a13335b03c7e02a95215",
                "requestBody": {
                    "description": "Pass User history  credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "format": "string",
                                        "example": "Jon Doe"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "string",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "text",
                                        "format": "string",
                                        "example": "1234567890AA"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "1234567890AA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout ",
                "description": "logout user !",
                "operationId": "0c6d22e05b701dd1f04b4421293a432f",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}